Work through adding `pub` because of warnings in a different way
I felt like the previous version was a bit vague and left the impression
that dead code warnings should be fixed by adding `pub` everywhere. What
I think it was *trying* to say was that because we WANT these functions
to be public for external folks to use, THAT'S why we're getting the
warnings and why pub is the fix for it.